+++ worker/src/worker.fallback.js
@@ -1,0 +1,16 @@
+/**
+ * Service worker browser client
+ * @package     GZip Plugin
+ * @copyright   Copyright (C) 2005 - 2018 Thierry Bela.
+ *
+ * dual licensed
+ *
+ * @license     LGPL v3
+ * @license     MIT License
+ */
+
+// @ts-check
+// placeholder for background sync task fallback
+
+import './sync/sw.sync.fallback.js';
+import './expiration/browser.cleanup.js';